Started work on Generic Flexi 2.0. Refactoring of Modes.

+++ src/main/java/de/mossgrabers/bitwig/framework/daw/TransportImpl.java
@@ -1,5 +1,5 @@
 // Written by Jürgen Moßgraber - mossgrabers.de
-// (c) 2017-2018
+// (c) 2017-2019
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
 package de.mossgrabers.bitwig.framework.daw;
@@ -537,9 +537,9 @@ public class TransportImpl implements ITransport
 
     /** {@inheritDoc} */
     @Override
-    public void setCrossfade (final int value)
+    public void setCrossfade (final double value)
     {
-        this.transport.crossfade ().set (Integer.valueOf (value), Integer.valueOf (this.valueChanger.getUpperBound ()));
+        this.transport.crossfade ().set (Double.valueOf (value), Integer.valueOf (this.valueChanger.getUpperBound ()));
     }
