+++ src/main/java/de/mossgrabers/framework/daw/ITransport.java
@@ -1,5 +1,5 @@
 // Written by Jürgen Moßgraber - mossgrabers.de
-// (c) 2017-2018
+// (c) 2017-2019
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
 package de.mossgrabers.framework.daw;
@@ -427,7 +427,7 @@ public interface ITransport extends ObserverManagement
      *
      * @param value The value
      */
-    void setCrossfade (final int value);
+    void setCrossfade (final double value);
 
 
     /**
